@@ -11,19 +11,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^13.3.5",
-    "@angular/animations": "~13.3.6",
+    "@angular-devkit/schematics": "^14.2.10",
+    "@angular/animations": "^14.2.12",
     "@angular/cdk": "^10.2.7",
-    "@angular/cli": "^13.3.5",
-    "@angular/common": "~13.3.6",
-    "@angular/compiler": "~13.3.6",
-    "@angular/core": "^13.3.6",
-    "@angular/forms": "~13.3.6",
-    "@angular/localize": "^13.3.6",
+    "@angular/cli": "^14.2.10",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/localize": "^14.2.12",
     "@angular/material": "^10.2.7",
-    "@angular/platform-browser": "~13.3.6",
-    "@angular/platform-browser-dynamic": "~13.3.6",
-    "@angular/router": "~13.3.6",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@ctrl/ngx-codemirror": "4.1.1",
     "@sentry/angular": "^6.3.6",
     "@sentry/tracing": "^6.3.6",
@@ -39,9 +39,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.5",
-    "@angular/compiler-cli": "~13.3.6",
-    "@angular/language-service": "~13.3.6",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/language-service": "^14.2.12",
     "@babel/preset-env": "^7.11.5",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
